requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.3
distlib==0.3.1
Django==3.1.6
django-crispy-forms==1.11.2
django-js-asset==1.2.2
filelock==3.0.12
idna==3.2
multidict==5.1.0
Pillow==8.2.0
pipenv==2020.11.15
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
yarl==1.6.3
